+++ src/components/2017/page-book-review-more-than-everything-by-vanessa-foster/page-book-review-more-than-everything-by-vanessa-foster.tsx
@@ -71,10 +71,9 @@ export class PageBookReviewMoreThanEverythingByVanessaFoster {
               <p>Today, finding a good book, much less a great one, is a difficult task; they are few and far between. But if you&#8217;re anything like me, I think you&#8217;ll find that Vanessa Foster&#8217;s <em>More Than Everything</em> is at least more than many.</p>
 
             </ion-col>
-            <ion-col size-xs="12" size-sm="12" size-md="4" size-lg="4" size-xl="5">
-
+            <ion-col size-xs="12" size-sm="12" size-md="4" size-lg="4" size-xl="5" style={{textAlign:`center`}}>
+              <p><em>As an Amazon Associate<br/> I earn from <br/>qualifying purchases.</em></p>
               <iframe title=" More Than Everything by Vanessa Foster" style={{ width: `120px`, height: `240px` }} marginwidth="0" marginheight="0" scrolling="no" frameborder="0" src="https://ws-na.amazon-adsystem.com/widgets/q?ServiceVersion=20070822&OneJS=1&Operation=GetAdHtml&MarketPlace=US&source=ac&ref=qf_sp_asin_til&ad_type=product_link&tracking_id=burtecgrollc-20&marketplace=amazon&region=US&placement=0615780873&asins=0615780873&linkId=a8b77a2c7c69ed9cab112ba6d16dd9d9&show_border=false&link_opens_in_new_window=false&price_color=333333&title_color=0066c0&bg_color=ffffff"></iframe>
-
             </ion-col>
           </ion-row>
         </ion-grid>