+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.6]
+
+- Added new page: [/vietnam-war-reference-resources](https://codyburleson.com/vietnam-war-reference-resources)
+- Added new page: [/cage/d-1-3-platoon](https://codyburleson.com/cage/d-1-3-platoon), but remains hidden until I can get better identifications
+- Modified page: [/cage](https://codyburleson.com/cage); reordered photo items to be alphabetical, added link to [/vietnam-war-reference-resources](https://codyburleson.com/vietnam-war-reference-resources)
 
 ## [1.6.5]